fix: resolve PlaylistAllSongViewModel Koin injection crash and data loading bug
- Fix InstanceCreationException by refining Koin viewModel parameters in AppModule
- Fix bug in PlaylistAllSongViewModel where songs flow was not collected
- Correct state update logic to properly display songs in playlist
- Reorder ViewModel constructor parameters for better DI consistency

+++ app/src/main/java/com/example/musicplayer/HomeScreen/ui/HomeScreenViewMode.kt
@@ -86,7 +86,7 @@ class HomeScreenViewModel(private val repository: MusicRepository , private val 
             is HomeEvent.AddToPlaylist -> {
                 viewModelScope.launch {
                     playlistRepository.insertSongFromPlaylist(audioFile = event.Audio ,
-                        event.playlistId.toInt()
+                        event.playlistId
                     )
                 }
             }
